requirements.txt
pyTelegramBotAPI==4.12.0
python-telegram-bot==20.8
pandas==2.2.0
numpy==1.26.4
psycopg2-binary==2.9.9
watchdog==3.0.0
pytz==2024.1 
